% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccusage.R
\name{append_to_duckdb}
\alias{append_to_duckdb}
\title{Append usage data to DuckDB}
\usage{
append_to_duckdb(
  data,
  db_path = "inst/extdata/llm_usage.duckdb",
  table_name = "daily_usage"
)
}
\arguments{
\item{data}{Usage data tibble}

\item{db_path}{Path to DuckDB file}

\item{table_name}{Table name}
}
\description{
Append usage data to DuckDB
}
