% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiverse.R
\name{.empty_runiverse_checks_tibble}
\alias{.empty_runiverse_checks_tibble}
\title{Empty \code{\link[=runiverse_checks]{runiverse_checks()}} tibble with the correct schema}
\usage{
.empty_runiverse_checks_tibble()
}
\description{
Empty \code{\link[=runiverse_checks]{runiverse_checks()}} tibble with the correct schema
}
\keyword{internal}
