% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiverse.R
\name{.empty_runiverse_packages_tibble}
\alias{.empty_runiverse_packages_tibble}
\title{Empty \code{\link[=runiverse_packages]{runiverse_packages()}} tibble with the correct schema}
\usage{
.empty_runiverse_packages_tibble()
}
\description{
Empty \code{\link[=runiverse_packages]{runiverse_packages()}} tibble with the correct schema
}
\keyword{internal}
