% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccusage.R
\name{fetch_ccusage}
\alias{fetch_ccusage}
\title{Fetch ccusage data from command line}
\usage{
fetch_ccusage(
  type = c("daily", "weekly", "session", "blocks"),
  project_filter = NULL
)
}
\arguments{
\item{type}{One of "daily", "weekly", "session", "blocks"}

\item{project_filter}{Project name to filter (NULL for all)}
}
\value{
tibble of usage data
}
\description{
Fetch ccusage data from command line
}
