% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccusage.R
\name{find_activity_gaps}
\alias{find_activity_gaps}
\title{Identify gaps in activity}
\usage{
find_activity_gaps(daily_data)
}
\arguments{
\item{daily_data}{Daily usage tibble}
}
\value{
tibble with gap information
}
\description{
Identify gaps in activity
}
