% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiverse.R
\name{\%||\%}
\alias{\%||\%}
\title{\verb{\%||\%} — null-coalescing operator (local copy, no rlang import needed)}
\usage{
a \%||\% b
}
\description{
\verb{\%||\%} — null-coalescing operator (local copy, no rlang import needed)
}
\keyword{internal}
