% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccusage.R
\name{parse_ccusage_json}
\alias{parse_ccusage_json}
\title{Parse ccusage JSON output}
\usage{
parse_ccusage_json(json_data, project_filter = NULL)
}
\arguments{
\item{json_data}{Parsed JSON from ccusage}

\item{project_filter}{Project name to filter}
}
\value{
tibble of usage data
}
\description{
Parse ccusage JSON output
}
