% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccusage.R
\name{query_latest_usage}
\alias{query_latest_usage}
\title{Query latest usage data from DuckDB}
\usage{
query_latest_usage(
  db_path = "inst/extdata/llm_usage.duckdb",
  table_name = "daily_usage"
)
}
\arguments{
\item{db_path}{Path to DuckDB file}

\item{table_name}{Table name}
}
\value{
tibble of latest data
}
\description{
Query latest usage data from DuckDB
}
