% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roborev_plots.R
\docType{data}
\name{ROBOREV_PALETTE}
\alias{ROBOREV_PALETTE}
\title{Severity colour palette for roborev charts}
\format{
An object of class \code{character} of length 6.
}
\usage{
ROBOREV_PALETTE
}
\description{
Severity colour palette for roborev charts
}
\keyword{datasets}
