% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roborev_plots.R
\docType{data}
\name{ROBOREV_TIER_PALETTE}
\alias{ROBOREV_TIER_PALETTE}
\title{Tier colour palette for roborev loop table row colouring}
\format{
An object of class \code{character} of length 3.
}
\usage{
ROBOREV_TIER_PALETTE
}
\description{
Tier colour palette for roborev loop table row colouring
}
\keyword{datasets}
