% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_version}
\alias{schema_version}
\title{Return the current schema version string}
\usage{
schema_version()
}
\value{
A character string, currently \code{"v1"}.
}
\description{
Return the current schema version string
}
